got ldap authentication mostly working with remember me functionality.
