--- frontend/app/lib/auth/EloquentLdapAuthenticatedUserProvider.php
@@ -1,5 +1,0 @@
-<?php namespace Paperwork;
-use Illuminate\Auth\EloquentUserProvider;
-class EloquentLdapAuthenticatedUserProvider extends EloquentUserProvider{
-    
-}
